Refactor: restaurant list fetching 위치 변경
