+++ db.json
@@ -35,6 +35,13 @@
       "name": "도스타코스 선릉점",
       "description": "멕시칸 캐주얼 그릴",
       "category": "기타"
+    },
+    {
+      "id": "94fe",
+      "category": "한식",
+      "name": "asd",
+      "description": "asd",
+      "imgSrc": "/templates/category-korean.png"
     }
   ]
 }
